chore(sort-numbers): consolidate skin dropdown into main story
Moves the Demo Pink skin registration and dropdown control into the
existing SortNumbers Storybook via argTypes, replacing the standalone
SkinHarness-based story. Keeps skin switching reachable from every
story (Consecutive, Distractors*, LevelMode*, etc.) without needing a
separate entry.
